+++ src/main/java/org/mmarini/wheelly/agents/TilesProcessor.java
@@ -27,36 +27,32 @@ package org.mmarini.wheelly.agents;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.jetbrains.annotations.NotNull;
 import org.mmarini.Tuple2;
 import org.mmarini.wheelly.envs.ArraySignal;
 import org.mmarini.wheelly.envs.FloatSignalSpec;
 import org.mmarini.wheelly.envs.Signal;
 import org.mmarini.wheelly.envs.SignalSpec;
-import org.mmarini.yaml.Utils;
 import org.mmarini.yaml.schema.Locator;
-import org.mmarini.yaml.schema.Validator;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.ops.transforms.Transforms;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.function.UnaryOperator;
-import java.util.stream.LongStream;
-import java.util.stream.Stream;
+import java.util.stream.Collectors;
 
 import static java.lang.Math.ceil;
 import static java.lang.Math.log;
-import static java.lang.String.format;
-import static org.mmarini.yaml.schema.Validator.*;
+import static org.mmarini.wheelly.agents.PartitionProcessor.*;
 
 /**
  * The processor creates a tile coding signal from float signals
  */
 public interface TilesProcessor {
-    Validator TILE__SPEC = objectPropertiesRequired(Map.of(
-            "name", string(),
-            "inputs", objectAdditionalProperties(positiveInteger())
-    ), List.of("name", "inputs"));
 
     static long computeInSpaceDim(Map<String, SignalSpec> inSpec, Map<String, Integer> inputs) {
         return inputs.keySet().stream()
@@ -74,94 +70,56 @@ public interface TilesProcessor {
      * @param inSpec  the input specification
      */
     static InputProcessor create(JsonNode root, Locator locator, Map<String, SignalSpec> inSpec) {
-        TILE__SPEC.apply(locator).accept(root);
-        String name1 = locator.path("name").getNode(root).asText();
-        Map<String, Integer> inputs = locator.path("inputs").propertyNames(root)
-                .map(t -> t.setV2(t._2.getNode(root).asInt()))
-                .collect(Tuple2.toMap());
+        TILE_SPEC.apply(locator).accept(root);
+        String outName = locator.path("name").getNode(root).asText();
+        List<Tuple2<String, Long>> tilesInfo = createTilesInfo(root, locator);
         // Validate inputs
-        for (String name : inputs.keySet()) {
-            SignalSpec spec = inSpec.get(name);
-            if (spec == null) {
-                throw new IllegalArgumentException(format(
-                        "Input \"%s\" undefined", name
-                ));
-            }
-            if (!(spec instanceof FloatSignalSpec)) {
-                throw new IllegalArgumentException(format(
-                        "Input %s must be %s (%s)", name, FloatSignalSpec.class.getName(), spec.getClass().getName()
-                ));
-            }
-        }
+        validateTiles(inSpec, tilesInfo);
+        // Creates processor json spec
+        ObjectNode jsonNode = createJsonNode(outName, tilesInfo);
 
         // Creates processor json spec
-        ObjectNode spec = Utils.objectMapper.createObjectNode();
-        spec.put("class", TilesProcessor.class.getName());
-        spec.put("name", name1);
-        ObjectNode inputsNode = Utils.objectMapper.createObjectNode();
-        inputs.forEach(inputsNode::put);
-        spec.set("inputs", inputsNode);
+        long[] numTilesByDim = computeNumTilesByDim(inSpec, tilesInfo);
 
         // Creates output spec
-        Map<String, SignalSpec> outputSpec1 = createSpec(inSpec, name1, inputs);
+        Map<String, SignalSpec> outputSpec = createSpec(inSpec, outName, numTilesByDim);
 
         // Creates the encode function
-        UnaryOperator<Map<String, Signal>> encode1 = createEncoder(inSpec, name1, inputs);
+        UnaryOperator<Map<String, Signal>> encode1 = createEncoder(inSpec, outName, tilesInfo, numTilesByDim);
 
-        return new InputProcessor(encode1, outputSpec1, spec);
+        return new InputProcessor(encode1, outputSpec, jsonNode);
     }
 
     /**
      * Creates the encode function
      *
      * @param inSpec   input specification
-     * @param name1    the signal name
+     * @param outName  the signal name
      * @param numTiles the num of tiles
      */
-    static UnaryOperator<Map<String, Signal>> createEncoder(Map<String, SignalSpec> inSpec, String name1, Map<String, Integer> numTiles) {
-        List<String> names = new ArrayList<>(numTiles.keySet());
-        Collections.sort(names);
-        long[] outSpaceSizes = names.stream()
-                .flatMapToLong(name -> {
-                    long nTiles = numTiles.get(name);
-                    long nDims = inSpec.get(name).getSize();
-                    return LongStream.range(0, nDims).map(ignored -> nTiles + 1);
-                })
-                .toArray();
-        UnaryOperator<INDArray> encoder = createTileEncoder(outSpaceSizes);
-        Map<String, UnaryOperator<INDArray>> normalizers = Tuple2.stream(numTiles)
-                .map(t -> t.setV2(normalize((FloatSignalSpec) inSpec.get(t._1))))
-                .collect(Tuple2.toMap());
+    static UnaryOperator<Map<String, Signal>> createEncoder(Map<String, SignalSpec> inSpec, String outName, List<Tuple2<String, Long>> numTiles, long[] numTilesByDim) {
+        UnaryOperator<INDArray> encoder = createTileEncoder(numTilesByDim);
+        List<Tuple2<String, UnaryOperator<INDArray>>> normalizers = createNormalizers(inSpec, numTiles);
 
-        UnaryOperator<Map<String, Signal>> f = (Map<String, Signal> x) -> {
-            INDArray[] y2 = names.stream()
-                    .map(name -> {
-                        INDArray y0 = x.get(name).toINDArray();
-                        INDArray y1 = normalizers.get(name).apply(y0);
-                        return y1;
-                    }).toArray(INDArray[]::new);
+        return (Map<String, Signal> x) -> {
+            INDArray[] y2 = normalizers.stream()
+                    .map(t -> t._2.apply(x.get(t._1).toINDArray()))
+                    .toArray(INDArray[]::new);
             INDArray y3 = Nd4j.hstack(y2);
             INDArray y4 = encoder.apply(y3);
             Map<String, Signal> result = new HashMap<>(x);
-            result.put(name1, new ArraySignal(y4));
+            result.put(outName, new ArraySignal(y4));
             return result;
         };
-        return f;
     }
 
-    /**
-     * Returns the tile specs (num. dimension, num_tiles + 1)
-     *
-     * @param inSpec   the input specifications
-     * @param numTiles the inputs num of tiles
-     */
-    static Stream<Tuple2<String, long[]>> createInputTileSpecs(Map<String, SignalSpec> inSpec, Map<String, Integer> numTiles) {
-        return Tuple2.stream(numTiles)
+    @NotNull
+    static List<Tuple2<String, UnaryOperator<INDArray>>> createNormalizers(Map<String, SignalSpec> inSpec, List<Tuple2<String, Long>> numTiles) {
+        return numTiles.stream()
                 .map(t -> {
-                    long n = inSpec.get(t._1).getSize();
-                    long m = t._2 + 1;
-                    return t.setV2(new long[]{n, m});
-                });
+                    UnaryOperator<INDArray> fn = normalize((FloatSignalSpec) inSpec.get(t._1), t._2);
+                    return t.setV2(fn);
+                }).collect(Collectors.toList());
     }
 
     /**
@@ -184,18 +142,14 @@ public interface TilesProcessor {
     /**
      * Returns the dimension
      *
-     * @param inSpec   the input specifications
-     * @param name     name of signal
-     * @param numTiles the number of tiles
+     * @param inSpec        the input specifications
+     * @param name          name of signal
+     * @param numTilesByDim the number of tiles by dimension
      */
-    static Map<String, SignalSpec> createSpec(Map<String, SignalSpec> inSpec, String name, Map<String, Integer> numTiles) {
+    static Map<String, SignalSpec> createSpec(Map<String, SignalSpec> inSpec, String name, long[] numTilesByDim) {
         // Compute input space dimension
-        long k = computeInSpaceDim(inSpec, numTiles);
-        long n = numTiling(k);
-        long numFeatures = createInputTileSpecs(inSpec, numTiles)
-                .map(Tuple2::getV2)
-                .mapToLong(x -> x[0] * x[1])
-                .reduce(1, (a, b) -> a * b) * n;
+        long n = numTiling(numTilesByDim.length);
+        long numFeatures = Arrays.stream(numTilesByDim).map(a -> a + 1).reduce(1L, (a, b) -> a * b) * n;
         Map<String, SignalSpec> result = new HashMap<>(inSpec);
         result.put(name, new FloatSignalSpec(new long[]{numFeatures}, 0, 1));
         return result;
@@ -208,14 +162,16 @@ public interface TilesProcessor {
      */
     static UnaryOperator<INDArray> createTileEncoder(long[] numTiles) {
         long n = numTiling(numTiles.length);
+        INDArray maxIndices = Nd4j.createFromArray(numTiles);
         long[] featuresShape = new long[numTiles.length + 1];
-        System.arraycopy(numTiles, 0, featuresShape, 1, numTiles.length);
         featuresShape[0] = n;
-        INDArray scale = Nd4j.createFromArray(new long[][]{numTiles}).subi(1);
+        for (int i = 0; i < numTiles.length; ++i) {
+            featuresShape[i + 1] = numTiles[i] + 1;
+        }
         INDArray offsets = createOffsets(numTiles.length);
         INDArray tilingIndices = Nd4j.arange(0, n).reshape(new long[]{n, 1});
         return x -> {
-            INDArray tilesIndices = Transforms.floor(x.mul(scale).add(offsets));
+            INDArray tilesIndices = Transforms.min(Transforms.max(Transforms.floor(x.add(offsets)), 0), maxIndices);
             INDArray indices = Nd4j.hstack(tilingIndices, tilesIndices);
             INDArray result = Nd4j.zeros(featuresShape);
             for (int i = 0; i < n; i++) {
@@ -227,15 +183,17 @@ public interface TilesProcessor {
     }
 
     /**
-     * Returns the normalizer function for a given signal spec
+     * Returns the normalized signal of x
+     * The output values are within (0, numValues - 1) range
      *
-     * @param spec the input signals specification
+     * @param inSpec    the input specification
+     * @param numValues the number of output value
      */
-    static UnaryOperator<INDArray> normalize(FloatSignalSpec spec) {
-        float min = spec.getMinValue();
-        float max = spec.getMaxValue();
-        float scale = 1 / (max - min);
-        return x -> Transforms.min(Transforms.max(x, min), max).sub(min).mul(scale);
+    static UnaryOperator<INDArray> normalize(FloatSignalSpec inSpec, long numValues) {
+        float minValue = inSpec.getMinValue();
+        float scale = (inSpec.getMaxValue() - minValue) / numValues;
+        return x ->
+                x.sub(minValue).mul(scale);
     }
 
     /**